HaploTraitR: Haplotype-Based Trait Analysis for Plant Breeding

HaploTraitR turns GWAS hits into haplotype blocks (SNPs linked by LD), tests how each haplotype affects a trait, and reports the superior haplotypes, diagnostic markers and candidate parents in plots and a formatted Excel workbook.

Details

The quickest way to use the package is [run_haplotraitr()]. The individual steps ([read_gwas_file()], [readHapmap()], [getDistClusters()], ...) remain available for custom workflows.

See also

Author

Maintainer: Alsamman Alsamman aalsamman100@gmail.com

Other contributors:

Tamara Ortiz tamara.ortizru@gmail.com [contributor]

Khaled Helmy khaledhelmy444@gmail.com [contributor]

Zakaria Kehel z.kehel@cgiar.org [contributor]

ICARDA [copyright holder]